requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
motor==3.6.0
pydantic==2.9.2
passlib==1.7.4
python-jose==3.3.0
pymongo==4.9.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.12
bcrypt===4.2.0
mongoengine===0.29.1
apscheduler==3.10.4